move semaphores logic to sync middleware
